Fix format linux/osx compat for 64bit fmt strings

+++ src/s3.c
@@ -1605,9 +1605,9 @@ static S3Status listPartsCallback(int isTruncated,
             strftime(timebuf, sizeof(timebuf), "%Y-%m-%dT%H:%M:%SZ",
                      gmtime(&t));
             printf("%-30s", timebuf);
-            printf("%-15llu", (unsigned long long) part->partNumber);
+            printf("%-15" PRIu64, part->partNumber);
             printf("%-45s", part->eTag);
-            printf("%-15llu\n", (unsigned long long) part->size);
+            printf("%-15" PRIu64 "\n", part->size);
         }
     }
